+++ uui-core/src/services/ApiContext.ts
@@ -369,12 +369,18 @@ export class ApiContext extends BaseContext implements IApiContext {
 
             xhr.onreadystatechange = () => {
                 if (xhr.readyState !== 4) return;
+                let response;
+                try {
+                    response = JSON.parse(xhr.response);
+                } catch {
+                    reject({ error: { isError: true, message: 'File upload error' } });
+                }
                 if (!new RegExp('^2[0-9][0-9]').test(xhr.status.toString())) {
-                    reject({ error: { isError: true, message: xhr.response && JSON.parse(xhr.response)?.error?.message } });
+                    reject({ error: { isError: true, message: response?.error?.message } });
                 }
 
                 removeAllListeners();
-                resolve((xhr.response && { ...JSON.parse(xhr.response) }) || null);
+                resolve(response);
             };
             xhr.send(formData);
         });